+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [2.5.2] 2024-07-17 (dev)
+## [2.5.2] 2024-07-18 (dev)
 ### Added
 - New {ifuser id|username}{/ifuser} tag.
 ### Updated
+- {categories0} tag now shows hidden categories if role has moodle/category:viewhiddencategories.
+- {categories0menu} tag now shows hidden categories if role has moodle/category:viewhiddencategories.
 - Fixed a couple of PHP 5.6 compatibility issues. Note that unit tests are still only compatible with PHP 7.1 and later.
 
 ## [2.5.1] 2024-05-01